Basic hotbar
Finished the basic "model" of a hotbar. It:
- generates the specified number of slots
- you can select the hotbar slots via number keys on the keyboard
- selected slot gets highlighted (for now it uses a RarityBorder as the highlight and it can be changed in the future)

As of this commit, adding items and all that is not really done yet, but given that the only 2 items we'll have is sword and health potion, it will be done soon

p.s. i assume the potion will be stackable, like you'll be able to hold up to 3 or something, so i'm gonna do that too